generalize wheel implementation to allow for sieve optimizations
By allowing the Wheel type to store a factor, we are able to store Wheels in our BinaryHeap to drastically
reduce the number of times we have the rearrange the heap to keep it as a valid heap. This results in a 4-5x
increase in throughput for the find() method.

This commit also replaces Wheel30 with a more general Wheel that is not tied to a specific choice of intiial
primes. By keeping everything relating to the choice of initial primes in a few constants near the top of the
file, we are able to easily swap out different sized Wheels. E.g., if we wanted to also filter out all multiples
of 7, all we would need to do is 1) rename WHEEL_235 to WHEEL_2357, 2) replace WHEEL_2357 with an array
consisting of all the numbers less than 2 * 3 * 5 * 7 that are coprime with 2, 3, 5, and 7, and 3) update
WHEEL_LENGTH to be the length of the resulting array. At the cost of a small amount of memory, this would result
in, theoretically, a 14% speedup, since 1/7 of all composites are multiples of 7.